[fix] Fix check condition to get key state iterator
To get key state iterator of a given frameIdx is to find the first iterator
of state which its index is equal to frameIdx in the list of LogStates.

Upper_bound return the first iterator itState which its index is greater than frameIdx.
So we check itState and itPrevState to choose the one which is closer to frameIdx