[feat] Extract features from horizontal neighborhood
For now, only horizontal neighborhood is extracted.
The features extracted lookalike SSKE features, except their technic of extraction is different, using the pros of Vertex Map structure. So the space gap and the depth gap are computed differently but applying the same theory.
The feature of angle is now the cosine of the angle instead of the absolute sine. We then decided to create a threshold that could be used for cosine in DSSKE and sine in SSKE : the angle threshold (one for planes and one for edges). See desc MR !324 for further explanation.