[PV][fix] Initialize slam with pose and maps
For the slam initialization, one can set initial pose only at the beginning
of the slam. If number of log states is greater than 1, a warning is displayed.
* A button "Set initial SLAM" is added in PV.
* A setter for TworldInit is added in SLAM.

If the user wants to set initial pose during the SLAM, it is equivalent to transform the
reference frame of the LiDAR trajectory. In this case, the initial pose should
be set with SetInitialPose (which calls TransformOdom) function.