[PV][refact] Add transform tree category in the interface
The previous group "Base to LiDAR transform" becomes "Transform tree".
It is used to set transforms for SLAM. A pose is provided by a file .mat
which contains a 4*4 transformation matrix.
The way to set pose with 6D vector with euler angle is removed from the
interface because of the different conventions of euler angle.