[ROS][refact] Manage abnormal time order
Previously, lidar conversion node considers the pointcloud from sensor is always
assembled in the order. So a frame duration is computed by time of the first and
the last point.

Sometimes pointcloud from the sensor driver can have an abnormal time order
due to the packet loss. Now the duration is computed by the min and max value
of time in a pointcloud.